requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langgraph>=0.0.30
streamlit>=1.30.0
python-dotenv>=1.0.0
python-docx>=1.0.0
diff-match-patch>=20230430
chromadb>=0.4.22
arxiv>=2.1.0
pypdf>=4.0.0
requests>=2.31.0
openai>=1.12.0
python-multipart>=0.0.9
